Rename EmptyCriteria namespace in tests
